refactor(replication): align ddl apply semantics and widen ddl masks to bigint

+++ src/include/catalog/pg_publication.h
@@ -56,7 +56,7 @@ CATALOG(pg_publication,6104,PublicationRelationId)
 	bool		pubviaroot;
 
 	/* DDL classes published by this publication */
-	int32		pubddl BKI_DEFAULT(0);
+	int64		pubddl BKI_DEFAULT(0);
 } FormData_pg_publication;
 
 /* ----------------
@@ -106,7 +106,7 @@ typedef struct Publication
 	char	   *name;
 	bool		alltables;
 	bool		pubviaroot;
-	int32		pubddl;
+	int64		pubddl;
 	PublicationActions pubactions;
 } Publication;
